Move EquatableAnalyzer to ApiDesignGuidelines analyzer.
