+++ src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/Microsoft.ApiDesignGuidelines.CSharp.Analyzers.csproj
@@ -134,8 +134,6 @@
     <Compile Include="CSharpOverloadOperatorEqualsOnOverridingValueTypeEquals.Fixer.cs" />
     <Compile Include="CSharpPassSystemUriObjectsInsteadOfStrings.cs" />
     <Compile Include="CSharpPassSystemUriObjectsInsteadOfStrings.Fixer.cs" />
-    <Compile Include="CSharpImplementIEquatableWhenOverridingObjectEquals.cs" />
-    <Compile Include="CSharpImplementIEquatableWhenOverridingObjectEquals.Fixer.cs" />
     <Compile Include="CSharpCancellationTokenParametersMustComeLast.cs" />
     <Compile Include="CSharpCancellationTokenParametersMustComeLast.Fixer.cs" />
     <Compile Include="CSharpDoNotDeclareStaticMembersOnGenericTypes.cs" />
