--- src/Microsoft.ApiDesignGuidelines.Analyzers/UnitTests/ImplementIEquatableWhenOverridingObjectEqualsTests.Fixer.cs
@@ -1,33 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Test.Utilities;
-using Microsoft.CodeAnalysis.UnitTests;
-using Xunit;
-
-namespace Microsoft.ApiDesignGuidelines.Analyzers.UnitTests
-{
-    public class ImplementIEquatableWhenOverridingObjectEqualsFixerTests : CodeFixTestBase
-    {
-        protected override DiagnosticAnalyzer GetBasicDiagnosticAnalyzer()
-        {
-            return new BasicImplementIEquatableWhenOverridingObjectEqualsAnalyzer();
-        }
-
-        protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
-        {
-            return new CSharpImplementIEquatableWhenOverridingObjectEqualsAnalyzer();
-        }
-
-        protected override CodeFixProvider GetBasicCodeFixProvider()
-        {
-            return new BasicImplementIEquatableWhenOverridingObjectEqualsFixer();
-        }
-
-        protected override CodeFixProvider GetCSharpCodeFixProvider()
-        {
-            return new CSharpImplementIEquatableWhenOverridingObjectEqualsFixer();
-        }
-    }
-}
