--- src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/BasicImplementIEquatableWhenOverridingObjectEquals.Fixer.vb
@@ -1,19 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports System.Composition
-Imports System.Linq
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.CodeFixes
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace Microsoft.ApiDesignGuidelines.Analyzers     
-    ''' <summary>
-    ''' RS0008: Implement IEquatable when overriding Object.Equals
-    ''' </summary>
-    <ExportCodeFixProvider(LanguageNames.VisualBasic), [Shared]>
-    Public NotInheritable Class BasicImplementIEquatableWhenOverridingObjectEqualsFixer
-        Inherits ImplementIEquatableWhenOverridingObjectEqualsFixer 
-
-    End Class
-End Namespace
