--- src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/BasicImplementIEquatableWhenOverridingObjectEquals.vb
@@ -1,17 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.Diagnostics
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace Microsoft.ApiDesignGuidelines.Analyzers   
-    ''' <summary>
-    ''' RS0008: Implement IEquatable when overriding Object.Equals
-    ''' </summary>
-    <DiagnosticAnalyzer(LanguageNames.VisualBasic)>
-    Public NotInheritable Class BasicImplementIEquatableWhenOverridingObjectEqualsAnalyzer
-        Inherits ImplementIEquatableWhenOverridingObjectEqualsAnalyzer
-
-    End Class
-End Namespace
