--- src/Roslyn.Diagnostics.Analyzers/Core/OverrideObjectEqualsObjectWhenImplementingIequatable.Fixer.cs
@@ -1,36 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Generic;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Editing;
-using System.Collections.Immutable;
-using System.Composition;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Roslyn.Diagnostics.Analyzers
-{                              
-    /// <summary>
-    /// RS0009: Override Object.Equals(object) when implementing Iequatable
-    /// </summary>
-    public abstract class OverrideObjectEqualsObjectWhenImplementingIequatableFixer : CodeFixProvider
-    {
-        public sealed override ImmutableArray<string> FixableDiagnosticIds => ImmutableArray.Create(OverrideObjectEqualsObjectWhenImplementingIequatableAnalyzer.RuleId);
-
-        public sealed override FixAllProvider GetFixAllProvider()
-        {
-            // See https://github.com/dotnet/roslyn/blob/master/docs/analyzers/FixAllProvider.md for more information on Fix All Providers
-            return WellKnownFixAllProviders.BatchFixer;
-        }
-
-        public sealed override async Task RegisterCodeFixesAsync(CodeFixContext context)
-        {                              
-            // This is to get rid of warning CS1998, please remove when implementing this analyzer
-            await new Task(() => { });
-            throw new NotImplementedException();
-        }
-    }
-}
