+++ src/Roslyn.Diagnostics.Analyzers/Core/Roslyn.Diagnostics.Analyzers.csproj
@@ -106,8 +106,6 @@
       <DependentUpon>RoslynDiagnosticsAnalyzersResources.resx</DependentUpon>
     </Compile>
     <Compile Include="DoNotUseGenericCodeActionCreateToCreateCodeAction.cs" />
-    <Compile Include="OverrideObjectEqualsObjectWhenImplementingIequatable.cs" />
-    <Compile Include="OverrideObjectEqualsObjectWhenImplementingIequatable.Fixer.cs" />
     <Compile Include="DoNotInvokeDiagnosticDescriptor.cs" />
     <Compile Include="SpecializedEnumerableCreationAnalyzer.cs" />
     <Compile Include="SymbolDeclaredEventMustBeGeneratedForSourceSymbols.cs" />
