--- src/Roslyn.Diagnostics.Analyzers/UnitTests/OverrideObjectEqualsObjectWhenImplementingIequatableTests.cs
@@ -1,23 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Test.Utilities;
-using Microsoft.CodeAnalysis.UnitTests;
-using Roslyn.Test.Utilities;
-using Xunit;
-
-namespace Roslyn.Diagnostics.Analyzers.UnitTests
-{
-    public class OverrideObjectEqualsObjectWhenImplementingIequatableTests : DiagnosticAnalyzerTestBase
-    {
-        protected override DiagnosticAnalyzer GetBasicDiagnosticAnalyzer()
-        {
-            return new BasicOverrideObjectEqualsObjectWhenImplementingIequatableAnalyzer();
-        }
-
-        protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
-        {
-            return new CSharpOverrideObjectEqualsObjectWhenImplementingIequatableAnalyzer();
-        }
-    }
-}
