--- src/Roslyn.Diagnostics.Analyzers/VisualBasic/BasicOverrideObjectEqualsObjectWhenImplementingIequatable.Fixer.vb
@@ -1,19 +1,0 @@
-' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports System.Composition
-Imports System.Linq
-Imports Microsoft.CodeAnalysis
-Imports Microsoft.CodeAnalysis.CodeFixes
-Imports Microsoft.CodeAnalysis.VisualBasic
-Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
-
-Namespace Roslyn.Diagnostics.Analyzers     
-    ''' <summary>
-    ''' RS0009: Override Object.Equals(object) when implementing Iequatable
-    ''' </summary>
-    <ExportCodeFixProvider(LanguageNames.VisualBasic), [Shared]>
-    Public NotInheritable Class BasicOverrideObjectEqualsObjectWhenImplementingIequatableFixer
-        Inherits OverrideObjectEqualsObjectWhenImplementingIequatableFixer 
-
-    End Class
-End Namespace
